add fixes

+++ app/src/main/java/am/te/myapplication/Model/Listing.java
@@ -122,7 +122,7 @@ public class Listing implements Comparable<Listing>{
 
     public static Listing getListingFromIntent(Intent intent) {
         Bundle extras = intent.getExtras();
-        Listing newListing = new Listing(extras.getString("Name"), extras.getInt("Price"), extras.getString("Additional"));
+        Listing newListing = new Listing(extras.getString("Name"), extras.getDouble("Price"), extras.getString("Additional"));
         return newListing;
     }
